add: custom hooks for password visibilty and toast message
